+++ Sawgle/Protocol/StackViewItemProtocol.swift
@@ -1,0 +1,11 @@
+//
+//  StackViewItemProtocol.swift
+//  Sawgle
+//
+//  Created by Milkyo on 28/07/2019.
+//  Copyright © 2019 Meo MacBook Pro. All rights reserved.
+//
+
+protocol StackViewItemProtocol {
+    func makeStackView()
+}
