+++ Sawgle/Protocol/UIViewItemProtocol.swift
@@ -6,21 +6,7 @@
 //  Copyright © 2019 Meo MacBook Pro. All rights reserved.
 //
 
-import Foundation
-
 protocol UIViewItemProtocol {
     func makeSubView()
     func makeSubViewConstraint()
-}
-
-protocol ActionProtocol {
-    func linkAction()
-}
-
-protocol StackViewItemProtocol {
-    func makeStackView()
-}
-
-protocol CellProtocol {
-    func registerCell()
 }
